# Reuse Kafka admin client for better performance