**Existing code with Kafka Admin client**

```
listOffsetsResult for topic pullRequestMergedEvents partition 0 and offset smallest took 9407ms

listOffsetsResult for topic pullRequestMergedEvents partition 1 and offset smallest took 10592ms

listOffsetsResult for topic pullRequestMergedEvents partition 2 and offset smallest took 9364ms

listOffsetsResult for topic pullRequestMergedEvents partition 3 and offset smallest took 7994ms
```

**With this PR**

```
listOffsetsResult for topic pullRequestMergedEvents partition 0 and offset smallest took 7669ms 

listOffsetsResult for topic pullRequestMergedEvents partition 1 and offset smallest took 6664ms

listOffsetsResult for topic pullRequestMergedEvents partition 2 and offset smallest took 6815ms 

listOffsetsResult for topic pullRequestMergedEvents partition 3 and offset smallest took 7628ms 
```

This will also lead to lower log generations. The logs are flooded with lines like these on every creation 

```
2025/06/20 03:16:18.903 WARN [AdminClientConfig] [foo_bar] The configuration 'key.deserializer' was supplied but isn't a known config.
2025/06/20 03:16:18.903 WARN [AdminClientConfig] [foo_bar] The configuration 'realtime.segment.flush.threshold.rows' was supplied but isn't a known config.
2025/06/20 03:16:18.903 WARN [AdminClientConfig] [foo_bar] The configuration 'value.deserializer' was supplied but isn't a known config.
```